Moved clock to submodule
