+++ src/dataflow/mod.rs
@@ -13,7 +13,7 @@ use crate::inputs::InputConfig;
 use crate::outputs::OutputConfig;
 use crate::pyo3_extensions::TdPyCallable;
 use crate::recovery::model::StepId;
-use crate::window::ClockConfig;
+use crate::window::clock::ClockConfig;
 use crate::window::WindowConfig;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
